requirements.txt
flask
flask-cors
sentence-transformers
scikit-learn
numpy
pandas
faiss-cpu
requests
python-dotenv